Avoid repeated Tesseract builds
Running autogen.sh creates a new configure file, then a new Makefile.in.
That updates the timestamp for tesseract.
With the dependency introduced in commit e1b30b95, that triggered a new run of
autogen.sh when make was run again and so on.

Replace tesseract/configure (both as a prerequisite to $(BIN)/tesseract and as
a target for the autogen recipe) with Makefile.in, the last file created.

Suggested-by: Robert Sachunsky <sachunsky@informatik.uni-leipzig.de>
Signed-off-by: Stefan Weil <sw@weilnetz.de>